refactor(temperature): make graph less stroboscopic
