+++ thorn/include/common/temperature.h
@@ -17,9 +17,9 @@
 #define TEMPERATURE_FAN_ON  50000
 #define TEMPERATURE_MAX     80000
 
-#define TEMPERATURE_CHECK_DELAY 5000000
+#define TEMPERATURE_CHECK_DELAY 2000
 
-#define TEMPERATURE_POINTS 100
+#define TEMPERATURE_POINTS 200
 
 #define TEMPERATURE_GRAPH_LOW  45000
 #define TEMPERATURE_GRAPH_HIGH 55000
