+++ thorn/src/kernel/entry.S
@@ -7,6 +7,7 @@
 	mov	x0, #\type
 	mrs	x1, esr_el1
 	mrs	x2, elr_el1
+	mrs	x3, far_el1
 	bl	show_invalid_entry_message
 	b	err_hang
 	.endm
